requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.9.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aom=3.6.0=h6a678d5_0
asttokens=3.0.0=pyhd8ed1ab_0
attrs=25.3.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.14.2=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli-python=1.0.9=py38h6a678d5_8
bypy=1.8.9=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.8.3=hbd8a1cb_0
cairo=1.18.4=h44eff21_0
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.2=pyhd8ed1ab_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.9=3
dav1d=1.2.1=h5eee18b_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
dill=0.4.0=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
expat=2.7.1=h6a678d5_0
fastjsonschema=2.21.2=pypi_0
ffmpeg=6.1.1=h2a67f75_3
filelock=3.13.1=py38h06a4308_0
fontconfig=2.14.1=h55d465d_3
freetype=2.13.3=h4a9f257_0
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.1.2=py38heeb90bb_0
graphite2=1.3.14=h295c915_1
h5py=3.11.0=pypi_0
harfbuzz=10.2.0=hdfddeaa_1
icu=73.1=h6a678d5_0
idna=3.7=py38h06a4308_0
imageio=2.35.1=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.0.0=h9e868ea_25371
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py38h06a4308_0
jpeg=9f=h5ce9db8_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.8.1=pyh31011fe_0
jupyterlab-pygments=0.3.0=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
leptonica=1.82.0=h42c8aad_2
lerc=4.0.0=h6a678d5_0
libarchive=3.8.1=h13d4d48_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.14.1.1=4
libcurand=10.3.10.19=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.84.2=h37c7471_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h5eee18b_1
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtheora=1.1.1=h7f8727e_3
libtiff=4.5.1=hffd6297_1
libuuid=1.41.5=h5eee18b_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.13.1=h6a678d5_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
libxml2=2.13.8=hfdd30dd_0
libzlib=1.3.1=hb25bd0a_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mistune=3.1.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py38h06a4308_0
multiprocess=0.70.18=pypi_0
nbclient=0.10.1=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h7934f7d_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=py38h06a4308_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.17=h5eee18b_0
packaging=25.0=pyh29332c3_1
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
pipreqs=0.5.0=pypi_0
pixman=0.46.4=h7934f7d_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_0
prompt_toolkit=3.0.48=hd8ed1ab_1
psutil=5.9.1=py38h0a891b7_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.20=he870216_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.0.0=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.3=hc2a1206_0
referencing=0.35.1=pypi_0
requests=2.32.3=py38h06a4308_0
requests-toolbelt=1.0.0=pypi_0
rpds-py=0.20.1=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=75.1.0=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.7=pypi_0
sqlite=3.50.2=hb25bd0a_1
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.3=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tesseract=5.2.0=hb0d2e87_3
tifffile=2023.7.10=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.15=h54e0aa7_0
torchaudio=2.0.0=py38_cu118
torchtriton=2.0.0=py38
torchvision=0.15.0=py38_cu118
tornado=6.1=py38h0a891b7_3
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.11.0=py38h06a4308_0
urllib3=2.2.3=py38h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
wheel=0.44.0=py38h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-libxext=1.3.6=h9b100fa_0
xorg-libxrender=0.9.12=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yarg=0.1.9=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.20.2=pypi_0
zlib=1.3.1=hb25bd0a_0
zstd=1.5.6=hc292b87_0
